Move V1Beta3 DatastoreSource into its own file
